Correcciones en los dao, repositories e intalacion de ksp
Hay un problema que no deja correr la app, son errores de sincronizacion y buildeo por haber hecho cambios en los daos

+++ app/src/main/java/LosPrimos/Durango/calculadoragastos/data/repositories/PresupuestoRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 
 class PresupuestoRepository(private val presupuestoDao: PresupuestoDao){
 
-    suspend fun insertarPresupuesto(presupuesto: Presupuesto): Int{
+    suspend fun insertarPresupuesto(presupuesto: Presupuesto): Long{
         return presupuestoDao.insertPresupuesto(presupuesto)
     }
